Fix curiosity in base reset
Prior to this commit, the base reset method was designed to favor
moves it hadn't seen before. This was an issue however because it was
searching its entire memory to see what moves it had done previously.
This meant it was discouraged from moves which had previously brought it
to the goal.

Now, it only looks at the moves it's done since the most recent time
it reached the goal.