Stop randomly generating the path to begin with
Prior to this commit, the agent was using the more intelligent reset
but was still finding it's initial path (before trimming) by walking
randomly. This meant that if the state machine generated had a very
isolated goal, there was a good chance the agent wouldn't find a path.

Now, the agent generates it's firt path using reset, so that it can
search for the goal more intelligently (by favoring moves it hasn't done)
so that we can avoid a situation where the agent never finds a path to the
goal.