Nearing Completion of SmartAgent
Since Last commit added/changed: Episode.java. Now contains A stateID
object instead of an integer. Added an equals method as well as a copy
constructor. StateID.java. New class that contains an int to hold the
stateID of an episode. This class contains a setter and getter method.
smartAgent.java. Implimented many of the empty header functions previously
listed in the run() method. All methods that are not in the run() method
are helper methods for one of the other methods that are called in run().