Various bug fixes
Fix bug where agent would not reset to state 1 when it hits the
goal state. Fix bug where agent would try to make a plan to the
lowest-numbered state with unknown transitions even if the current
state had unknown transitions. Hard-code transitions for the
goal state so the Agent doesn't think it needs to fill out those
transitions.