Fixed a few bugs in handling equivalent state hypotheses.
The maxMatchedState() method had an off by one error (or the code using it did...) and we removed the code that makes certain that two candidate equivalent states have at least one non-unknown state identical.  I don't think we can guarantee this.  We left the code in there commented out so we can put it back in again if this bites us in the ass.