Added test functionality and deleted old methods
Deleted old versions of methods that were no longer in use. Added an
instance variable that can hold an arraylist of a defined path for the
agent to take. If a boolean flag is thrown in selectNextCommand it will
allow the tester to bypass the random move generator and force the agent
to take a specific path.