Fixed table size test and added best path test
Fixed test that would check equivalency of the agent's size vs real state machine. Also
added test that checks the best path the agent comes up with and the real best path
to assume they are the same. The environment has numerous version paths so it currently
just uses the first (zeroith element).